chore: renomeia e move componentes usados em TabBarView's para widgets/tabs
